Reuse frame.
